rename settings package
