+++ apps/Android/MnnLlmChat/app/src/main/java/com/alibaba/mls/api/source/ModelSources.kt
@@ -3,7 +3,7 @@
 package com.alibaba.mls.api.source
 
 import com.alibaba.mls.api.ApplicationProvider
-import com.alibaba.mnnllm.android.settings.MainSettings.getDownloadProvider
+import com.alibaba.mnnllm.android.mainsettings.MainSettings.getDownloadProvider
 
 class ModelSources {
     enum class ModelSourceType {
