+++ apps/Android/MnnLlmChat/app/src/main/java/com/alibaba/mnnllm/android/mainsettings/MainSettings.kt
@@ -1,6 +1,6 @@
 // Created by ruoyi.sjd on 2025/2/28.
 // Copyright (c) 2024 Alibaba Group Holding Limited All rights reserved.
-package com.alibaba.mnnllm.android.settings
+package com.alibaba.mnnllm.android.mainsettings
 
 import android.content.Context
 import androidx.preference.PreferenceManager
