+++ apps/Android/MnnLlmChat/app/src/main/java/com/alibaba/mnnllm/android/mainsettings/MainSettingsFragment.kt
@@ -1,7 +1,7 @@
 // Created by ruoyi.sjd on 2025/2/28.
 // Copyright (c) 2024 Alibaba Group Holding Limited All rights reserved.
 
-package com.alibaba.mnnllm.android.settings
+package com.alibaba.mnnllm.android.mainsettings
 import android.os.Bundle
 import android.widget.Toast
 import androidx.preference.ListPreference
